LTID/requirements.txt
# Threat Intelligence Dashboard - Dependencies
# API Clients
requests>=2.31.0
vt-py>=0.18.0
OTXv2>=1.5.0

# Data Processing
pandas>=2.1.0
python-dotenv>=1.0.0

# Web Dashboard
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
jinja2>=3.1.0
slowapi>=0.1.9

# Grafana/Prometheus integration
prometheus-client>=0.19.0
